+++ geotrek/outdoor/translation.py
@@ -4,7 +4,7 @@ from modeltranslation.translator import translator, TranslationOptions
 
 
 class SiteTO(TranslationOptions):
-    fields = ('name', 'description', 'description_teaser', 'ambiance', 'advice', 'period') + (
+    fields = ('name', 'description', 'description_teaser', 'ambiance', 'accessibility', 'advice', 'period') + (
         ('published',) if settings.PUBLISHED_BY_LANG else tuple())
     fallback_undefined = {'published': None}