+++ geotrek/settings/base.py
@@ -643,6 +643,7 @@ TINYMCE_DEFAULT_CONFIG = {
     "toolbar": "bold italic forecolor | bullist numlist link media | "
                "undo redo | "
                "removeformat | code",
+    "paste_as_text": True
 }
 
 SYNC_RANDO_ROOT = os.path.join(VAR_DIR, 'data')
@@ -818,6 +819,8 @@ USE_BOOKLET_PDF = False
 HIDDEN_FORM_FIELDS = {}
 COLUMNS_LISTS = {}
 ENABLE_JOBS_COSTS_DETAILED_EXPORT = False
+
+ACCESSIBILITY_ATTACHMENTS_ENABLED = True
 
 # Override with prod/dev/tests/tests_nds settings
 ENV = os.getenv('ENV', 'prod')