+++ geotrek/trekking/serializers.py
@@ -226,14 +226,14 @@ class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
         model = trekking_models.Trek
         id_field = 'id'  # By default on this model it's topo_object = OneToOneField(parent_link=True)
         fields = ('id', 'departure', 'arrival', 'duration', 'duration_pretty',
-                  'description', 'description_teaser', 'networks', 'advice', 'equipment',
+                  'description', 'description_teaser', 'networks', 'advice', 'gear',
                   'ambiance', 'difficulty', 'information_desks', 'themes',
                   'labels', 'practice', 'accessibilities', 'accessibility_level',
                   'accessibility_signage', 'accessibility_slope', 'accessibility_covering', 'accessibility_exposure',
                   'accessibility_width', 'accessibility_advice',
                   'usages', 'access', 'route',
                   'public_transport', 'advised_parking', 'web_links',
-                  'disabled_infrastructure', 'parking_location', 'relationships',
+                  'accessibility_infrastructure', 'parking_location', 'relationships',
                   'points_reference', 'gpx', 'kml', 'source', 'portal',
                   'type2', 'category', 'structure', 'treks', 'reservation_id', 'reservation_system',
                   'children', 'parents', 'previous', 'next', 'ratings', 'ratings_description') + \