+++ geotrek/trekking/translation.py
@@ -9,7 +9,7 @@ from geotrek.trekking import models as trekking_models
 
 class TrekTO(TranslationOptions):
     fields = ('name', 'departure', 'arrival', 'description_teaser', 'description', 'ambiance', 'access',
-              'disabled_infrastructure', 'advice', 'equipment', 'accessibility_signage', 'accessibility_slope',
+              'accessibility_infrastructure', 'advice', 'gear', 'accessibility_signage', 'accessibility_slope',
               'accessibility_covering', 'accessibility_exposure', 'accessibility_width',
               'accessibility_advice', 'advised_parking', 'public_transport', 'ratings_description') + (
         ('published',) if settings.PUBLISHED_BY_LANG else tuple())