+++ geotrek/trekking/views.py
@@ -76,11 +76,11 @@ class TrekFormatList(MapEntityFormat, TrekList):
     mandatory_columns = ['id', 'name']
     default_extra_columns = [
         'eid', 'eid2', 'structure', 'departure', 'arrival', 'duration', 'duration_pretty', 'description',
-        'description_teaser', 'networks', 'advice', 'equipment', 'ambiance', 'difficulty', 'information_desks',
+        'description_teaser', 'networks', 'advice', 'gear', 'ambiance', 'difficulty', 'information_desks',
         'themes', 'practice', 'ratings', 'accessibilities', 'accessibility_advice', 'accessibility_covering',
         'accessibility_exposure', 'accessibility_level', 'accessibility_signage', 'accessibility_slope',
         'accessibility_width', 'access', 'route', 'public_transport', 'advised_parking', 'web_links', 'labels',
-        'disabled_infrastructure', 'parking_location', 'points_reference',
+        'accessibility_infrastructure', 'parking_location', 'points_reference',
         'related', 'children', 'parents', 'pois', 'review', 'published',
         'publication_date', 'date_insert', 'date_update',
         'cities', 'districts', 'areas', 'source', 'portal', 'length_2d', 'uuid',
@@ -509,7 +509,7 @@ class TrekServiceViewSet(viewsets.ModelViewSet):
 
 # Translations for public PDF
 translation.gettext_noop("Advices")
-translation.gettext_noop("Equipments")
+translation.gettext_noop("Gear")
 translation.gettext_noop("All useful information")
 translation.gettext_noop("Altimetric profile")
 translation.gettext_noop("Attribution")